pass globals to kubecmd
