+++ pkg/synopsysctl/utils.go
@@ -23,6 +23,8 @@ package synopsysctl
 
 import (
 	"fmt"
+	"os"
+	"os/exec"
 
 	horizonapi "github.com/blackducksoftware/horizon/pkg/api"
 	horizoncomponents "github.com/blackducksoftware/horizon/pkg/components"
@@ -72,7 +74,7 @@ func setResourceClients() {
 	if err != nil {
 		log.Errorf("error creating the Alert Clientset: %s", err)
 	}
-	kube, openshift = operatorutil.DetermineClusterClients(restconfig)
+	kube, openshift = DetermineClusterClients(restconfig)
 }
 
 // getKubeClient gets the kubernetes client
@@ -108,4 +110,115 @@ func GetOperatorNamespace() (string, error) {
 		return "", fmt.Errorf("failed to get synopsys operator namespace due to %+v", err)
 	}
 	return crb.Subjects[0].Namespace, nil
+}
+
+// DetermineClusterClients returns bool values for which client
+// to use. They will never both be true
+func DetermineClusterClients(restConfig *rest.Config) (kube, openshift bool) {
+	openshift = false
+	kube = false
+
+	kubectlPath := false
+	ocPath := false
+	_, exists := exec.LookPath("kubectl")
+	if exists == nil {
+		kubectlPath = true
+	}
+	_, ocexists := exec.LookPath("oc")
+	if ocexists == nil {
+		ocPath = true
+	}
+
+	// Add Openshift rules
+	openshiftTest := false
+	routeClient := operatorutil.GetRouteClient(restConfig) // kube doesn't have a route client but openshift does
+	if routeClient != nil {
+		openshiftTest = true
+	}
+
+	if ocPath && openshiftTest { // if oc exists and the cluster is openshift
+		log.Debugf("oc exists and the cluster is openshift")
+		return false, true
+	}
+	if kubectlPath && !openshiftTest { // if kubectl exists and it isn't openshift
+		log.Debugf("kubectl exists and it isn't openshift")
+		return true, false
+	}
+	if kubectlPath && !ocPath && openshiftTest { // if kubectl exists, oc doesn't exist, and it is openshift
+		log.Debugf("kubectl exists, oc doesn't exist, and it is openshift")
+		return true, false
+	}
+	if ocPath && !kubectlPath && !openshiftTest { // If oc exists, kubectl doesn't exist, and it isn't openshift
+		log.Debugf("oc exists, kubectl doesn't exist, and it isn't openshift")
+		return false, true
+	}
+	return false, false // neither client exists
+}
+
+// RunKubeCmd is a simple wrapper to oc/kubectl exec that captures output.
+// TODO consider replacing w/ go api but not crucial for now.
+func RunKubeCmd(restConfig *rest.Config, kube bool, openshift bool, args ...string) (string, error) {
+	var cmd2 *exec.Cmd
+
+	// cluster-info in kube doesnt seem to be in
+	// some versions of oc, but status is.
+	// double check this.
+	if args[0] == "cluster-info" && openshift {
+		args[0] = "status"
+	}
+	// add global flags: insecure-skip-tls-verify and --kubeconfig
+	if insecureSkipTLSVerify == true {
+		args = append([]string{fmt.Sprintf("--insecure-skip-tls-verify=%t", insecureSkipTLSVerify)}, args...)
+	}
+	if kubeconfig != "" {
+		args = append([]string{fmt.Sprintf("--kubeconfig=%s", kubeconfig)}, args...)
+	}
+	if openshift {
+		cmd2 = exec.Command("oc", args...)
+		log.Debugf("Command: %+v", cmd2.Args)
+	} else if kube {
+		cmd2 = exec.Command("kubectl", args...)
+		log.Debugf("Command: %+v", cmd2.Args)
+	} else {
+		return "", fmt.Errorf("Could not determine if openshift or kube")
+	}
+	stdoutErr, err := cmd2.CombinedOutput()
+	if err != nil {
+		return string(stdoutErr), err
+	}
+	//time.Sleep(1 * time.Second) TODO why did Jay put this here???
+	return string(stdoutErr), nil
+}
+
+// RunKubeEditorCmd is a wrapper for oc/kubectl but redirects
+// input/output to the user - ex: let user control text editor
+func RunKubeEditorCmd(restConfig *rest.Config, kube bool, openshift bool, args ...string) error {
+	var cmd *exec.Cmd
+
+	// cluster-info in kube doesnt seem to be in
+	// some versions of oc, but status is.
+	// double check this.
+	if args[0] == "cluster-info" && openshift {
+		args[0] = "status"
+	}
+	// add global flags: insecure-skip-tls-verify and --kubeconfig
+	if insecureSkipTLSVerify == true {
+		args = append([]string{fmt.Sprintf("--insecure-skip-tls-verify=%t", insecureSkipTLSVerify)}, args...)
+	}
+	if kubeconfig != "" {
+		args = append([]string{fmt.Sprintf("--kubeconfig=%s", kubeconfig)}, args...)
+	}
+	if openshift {
+		cmd = exec.Command("oc", args...)
+	} else if kube {
+		cmd = exec.Command("kubectl", args...)
+	}
+	cmd.Stdin = os.Stdin
+	cmd.Stdout = os.Stdout
+	err := cmd.Run()
+	if err != nil {
+		return err
+	}
+	//time.Sleep(1 * time.Second) TODO why did Jay put this here???
+	return nil
 }
