--- pkg/util/kubecmd.go
@@ -1,130 +1,0 @@
-/*
-Copyright (C) 2019 Synopsys, Inc.
-
-Licensed to the Apache Software Foundation (ASF) under one
-or more contributor license agreements. See the NOTICE file
-distributed with this work for additional information
-regarding copyright ownership. The ASF licenses this file
-to you under the Apache License, Version 2.0 (the
-"License"); you may not use this file except in compliance
-with the License. You may obtain a copy of the License at
-
-http://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing,
-software distributed under the License is distributed on an
-"AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
-KIND, either express or implied. See the License for the
-specific language governing permissions and limitations
-under the License.
-*/
-
-package util
-
-import (
-	"fmt"
-	"os"
-	"os/exec"
-
-	log "github.com/sirupsen/logrus"
-
-	"k8s.io/client-go/rest"
-	//_ "k8s.io/client-go/plugin/pkg/client/auth/gcp"
-)
-
-// DetermineClusterClients returns bool values for which client
-// to use. They will never both be true
-func DetermineClusterClients(restConfig *rest.Config) (kube, openshift bool) {
-	openshift = false
-	kube = false
-
-	kubectlPath := false
-	ocPath := false
-	_, exists := exec.LookPath("kubectl")
-	if exists == nil {
-		kubectlPath = true
-	}
-	_, ocexists := exec.LookPath("oc")
-	if ocexists == nil {
-		ocPath = true
-	}
-
-	// Add Openshift rules
-	openshiftTest := false
-	routeClient := GetRouteClient(restConfig) // kube doesn't have a route client but openshift does
-	if routeClient != nil {
-		openshiftTest = true
-	}
-
-	if ocPath && openshiftTest { // if oc exists and the cluster is openshift
-		log.Debugf("oc exists and the cluster is openshift")
-		return false, true
-	}
-	if kubectlPath && !openshiftTest { // if kubectl exists and it isn't openshift
-		log.Debugf("kubectl exists and it isn't openshift")
-		return true, false
-	}
-	if kubectlPath && !ocPath && openshiftTest { // if kubectl exists, oc doesn't exist, and it is openshift
-		log.Debugf("kubectl exists, oc doesn't exist, and it is openshift")
-		return true, false
-	}
-	if ocPath && !kubectlPath && !openshiftTest { // If oc exists, kubectl doesn't exist, and it isn't openshift
-		log.Debugf("oc exists, kubectl doesn't exist, and it isn't openshift")
-		return false, true
-	}
-	return false, false // neither client exists
-}
-
-// RunKubeCmd is a simple wrapper to oc/kubectl exec that captures output.
-// TODO consider replacing w/ go api but not crucial for now.
-func RunKubeCmd(restConfig *rest.Config, kube bool, openshift bool, args ...string) (string, error) {
-	var cmd2 *exec.Cmd
-
-	// cluster-info in kube doesnt seem to be in
-	// some versions of oc, but status is.
-	// double check this.
-	if args[0] == "cluster-info" && openshift {
-		args[0] = "status"
-	}
-	if openshift {
-		cmd2 = exec.Command("oc", args...)
-		log.Debugf("Command: %+v", cmd2.Args)
-	} else if kube {
-		cmd2 = exec.Command("kubectl", args...)
-		log.Debugf("Command: %+v", cmd2.Args)
-	} else {
-		return "", fmt.Errorf("Could not determine if openshift or kube")
-	}
-	stdoutErr, err := cmd2.CombinedOutput()
-	if err != nil {
-		return string(stdoutErr), err
-	}
-	//time.Sleep(1 * time.Second) TODO why did Jay put this here???
-	return string(stdoutErr), nil
-}
-
-// RunKubeEditorCmd is a wrapper for oc/kubectl but redirects
-// input/output to the user - ex: let user control text editor
-func RunKubeEditorCmd(restConfig *rest.Config, kube bool, openshift bool, args ...string) error {
-	var cmd *exec.Cmd
-
-	// cluster-info in kube doesnt seem to be in
-	// some versions of oc, but status is.
-	// double check this.
-	if args[0] == "cluster-info" && openshift {
-		args[0] = "status"
-	}
-	if openshift {
-		cmd = exec.Command("oc", args...)
-	} else if kube {
-		cmd = exec.Command("kubectl", args...)
-	}
-	cmd.Stdin = os.Stdin
-	cmd.Stdout = os.Stdout
-	err := cmd.Run()
-	if err != nil {
-		return err
-	}
-	//time.Sleep(1 * time.Second) TODO why did Jay put this here???
-	return nil
-}
